+++ tests/testthat/test-betaSandwich-r-sq-beta-sandwich.R
@@ -16,6 +16,22 @@ lapply(
     }
     df <- nas1982
     object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = df)
+    ############################################################
+    # coverage #################################################
+    ############################################################
+    lm_process <- .ProcessLM(object)
+    .JacobianVechSigmaWRTTheta(
+      beta = lm_process$beta,
+      sigmacapx = lm_process$sigmacap[
+        2:lm_process$k,
+        2:lm_process$k,
+        drop = TRUE
+      ],
+      q = lm_process$q,
+      p = lm_process$p,
+      rsq = NULL
+    )
+    ############################################################
     r_sq <- summary(object)$r.squared
     adj <- summary(object)$adj.r.squared
     mvn <- RSqBetaSandwich(BetaN(object))