+++ NEWS.md
@@ -2,13 +2,13 @@
 
 Latest development version.
 
-* Added the `DiffBetaSandwich` and `RSqBetaSandwich` functions.
+* Added the `DiffBetaSandwich()` and the `RSqBetaSandwich()` functions.
 
 # betaSandwich 1.0.3
 
 ## Patch
 
-* Added the `BetaADF` function.
+* Added the `BetaADF()` function.
 
 # betaSandwich 1.0.2
 
@@ -20,7 +20,7 @@ Latest development version.
 
 ## Patch
 
-* Added `nas1982`.
+* Added the `nas1982` data set.
 * Minor edits to methods.
 
 # betaSandwich 1.0.0